micro-services/dl_service/requirements.txt
fastapi
pydantic
uvicorn
requests
newspaper3k
beautifulsoup4
urlextract
